Initial support for displaying text inside challenge_modal

+++ game/common/position.go
@@ -1,5 +1,5 @@
 package common
 
 type Position struct {
-	X, Y int
+	X, Y float64
 }
